+++ investment_local_horse_racing_crawler/scrapy/settings.py
@@ -27,6 +27,7 @@ HTTPCACHE_STORAGE = 'investment_local_horse_racing_crawler.scrapy.middlewares.S3
 
 SPIDER_CONTRACTS = {
     "investment_local_horse_racing_crawler.scrapy.contracts.ScheduleListContract": 10,
+    "investment_local_horse_racing_crawler.scrapy.contracts.OneDayRaceListContract": 10,
     "investment_local_horse_racing_crawler.scrapy.contracts.RaceRefundListContract": 10,
     "investment_local_horse_racing_crawler.scrapy.contracts.RaceDenmaContract": 10,
     "investment_local_horse_racing_crawler.scrapy.contracts.OddsWinContract": 10,